backend/pyproject.toml
[tool.poetry]
name = "double-agent-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "app"}]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
langchain-core = "^0.3.0"
langgraph = "^0.2.28"
langchain-openai = "^0.2.0"
python-dotenv = "^1.0.0"
authlib = "^1.3.0"
itsdangerous = "^2.2.0"
sqlmodel = {extras = ["asyncio"], version = "^0.0.27"}
psycopg2-binary = "^2.9.10"
alembic = "^1.16.5"
sqlalchemy = "^2.0.44"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "app.dev:main"

[tool.black]
line-length = 88
target-version = ['py311']